// src/models/Items/Item.ts
export type CreateItemRequest = {
  stockcode: string | null;
  shortDesc: string | null;
  stage: number | null;
  status: number | null;
  status_Factory: string;
  familyName: string | null;
  category: number | null;
  style: number | null;
  function: number | null;
  skuType: number | null;
  slType: number | null;
  supplier: number | null;
  baseCode: string | null;
  baseCode_MKT: string | null;
  primaryFinishId: number | null;
  createDate: Date | null;
  coordinator: number | null;
  drawingNum: string | null;
  pdCode: string | null;
  priority: number | null;
  introdate: Date | null;
  marketingDesigner: number | null;
  royaltyDesigner: number | null;
  brand: number | null;
  // attachment: string | null;
  // twp: string | null;
};